Moved db update methods to class methods in model.py. Hid favorite button on homepage. Removed images from db that had no colors recorded in new db.
