+++ model.py
@@ -1,6 +1,7 @@
 """ Models and database functions for paleta db. """
 
 from flask_sqlalchemy import SQLAlchemy
+from helpers import get_color_bin
 
 db = SQLAlchemy()
 
@@ -55,6 +56,25 @@ class Image(db.Model):
     # There is a relationship defined between Image and ImageColor in ImageColor.
     # Backref to ImageColor is "imagecolors"
 
+    @classmethod
+    def add_image_to_db(cls, file_name):
+        # Next, check if the image is already in the db 
+        image_in_db = cls.query.filter(cls.file_name==file_name).first()
+
+        if image_in_db:
+            # colors = image_in_db.colors
+            image_id = image_in_db.image_id
+        # If not, add the image to the db
+        else:
+            new_photo = cls(file_name=file_name)
+            db.session.add(new_photo)
+            db.session.commit()
+
+            image_id = new_photo.image_id
+
+        return image_id
+
+
 
 class Color(db.Model):
     """ Color details """
@@ -67,6 +87,24 @@ class Color(db.Model):
 
     # There is a relationship defined between Color and ImageColor in ImageColor.
     # Backref is "imagecolors"
+
+    @classmethod
+    def add_colors_to_db(cls, colors):
+        """ Takes a list of comma-separated hex values, and makes a record in 
+            Color if none existed previously """
+
+        for color in colors:
+            color_in_db = cls.query.filter(cls.color==color).first()
+            # Color not already in the table
+            if not color_in_db:
+                new_color = cls(color=color)
+                db.session.add(new_color)
+                db.session.commit()
+        return
+
+
+
+    
     
 
 class ImageColor(db.Model):
@@ -89,6 +127,30 @@ class ImageColor(db.Model):
     color = db.relationship("Color", backref=db.backref("imagecolors"), order_by=image_color_id)
     
 
+    @classmethod
+    def add_image_colors_to_db(cls, image_id, colors):
+
+        for color in colors:
+
+            color_record = Color.query.filter(Color.color==color).first()
+            color_id = color_record.color_id
+
+            has_image_color = cls.query.filter(cls.image_id==image_id, 
+                                               cls.color_id==color_id).first()
+
+            if not has_image_color:
+
+                color_bin = get_color_bin(color)
+
+                new_image_color = cls(image_id=image_id, 
+                                             color_id=color_id,
+                                             color_bin=color_bin)
+                print 'new_image_color', new_image_color
+                db.session.add(new_image_color)
+                db.session.commit()
+
+
+        return
 
 
 ################### Helper Functions ####################
